+++ src/main/java/nextstep/subway/domain/line/dto/LineResponse.java
@@ -1,25 +1,27 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.domain.line.dto;
 
-import nextstep.subway.domain.Station;
+import nextstep.subway.domain.line.Line;
 
 import java.time.LocalDateTime;
 
-public class StationResponse {
+public class LineResponse {
     private final Long id;
     private final String name;
+    private final String color;
     private final LocalDateTime createdDate;
     private final LocalDateTime modifiedDate;
 
-    private StationResponse(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+    private LineResponse(Long id, String name, String color, LocalDateTime createdDate, LocalDateTime modifiedDate) {
         this.id = id;
         this.name = name;
+        this.color = color;
         this.createdDate = createdDate;
         this.modifiedDate = modifiedDate;
     }
 
-    public static StationResponse from(Station station) {
-        return new StationResponse(station.getId(), station.getName(), station.getCreatedDate(),
-                station.getModifiedDate());
+    public static LineResponse from(Line savedLine) {
+        return new LineResponse(savedLine.getId(), savedLine.getName(), savedLine.getColor(),
+                savedLine.getCreatedDate(), savedLine.getModifiedDate());
     }
 
     public Long getId() {
@@ -28,6 +30,10 @@ public class StationResponse {
 
     public String getName() {
         return name;
+    }
+
+    public String getColor() {
+        return color;
     }
 
     public LocalDateTime getCreatedDate() {
