+++ src/main/java/nextstep/subway/domain/section/SectionRepository.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.section;
 
 import org.springframework.data.jpa.repository.JpaRepository;
