+++ src/main/java/nextstep/subway/domain/station/StationRepository.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.station;
 
 import org.springframework.data.jpa.repository.JpaRepository;
