+++ src/main/java/nextstep/subway/domain/station/dto/StationResponse.java
@@ -1,27 +1,25 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.domain.station.dto;
 
-import nextstep.subway.domain.Line;
+import nextstep.subway.domain.station.Station;
 
 import java.time.LocalDateTime;
 
-public class LineResponse {
+public class StationResponse {
     private final Long id;
     private final String name;
-    private final String color;
     private final LocalDateTime createdDate;
     private final LocalDateTime modifiedDate;
 
-    private LineResponse(Long id, String name, String color, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+    private StationResponse(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
         this.id = id;
         this.name = name;
-        this.color = color;
         this.createdDate = createdDate;
         this.modifiedDate = modifiedDate;
     }
 
-    public static LineResponse from(Line savedLine) {
-        return new LineResponse(savedLine.getId(), savedLine.getName(), savedLine.getColor(),
-                savedLine.getCreatedDate(), savedLine.getModifiedDate());
+    public static StationResponse from(Station station) {
+        return new StationResponse(station.getId(), station.getName(), station.getCreatedDate(),
+                station.getModifiedDate());
     }
 
     public Long getId() {
@@ -30,10 +28,6 @@ public class LineResponse {
 
     public String getName() {
         return name;
-    }
-
-    public String getColor() {
-        return color;
     }
 
     public LocalDateTime getCreatedDate() {
