+++ src/main/java/nextstep/subway/handler/error/custom/BusinessException.java
@@ -1,6 +1,6 @@
 package nextstep.subway.handler.error.custom;
 
-public class BusinessException extends RuntimeException{
+public class BusinessException extends RuntimeException {
     private final ErrorCode errorCode;
 
     public BusinessException(ErrorCode errorCode) {
